@`README.md`; experimental versus trunk branch use.
	@`c/README.md`, @`cxx/README.md`, @`hooks/README.md`; ditto.
	@`sh/Macros.sh:SUSUWU_FAST_IS_PREVIEW()`; return true (until this commit, is false).

@`build.sh`; warn that this is the `experimental` branch, which now sets `-DSUSUWU_EXPERIMENTAL`.
	Is followup to: commit 59b533e5374d4a6f9d04f9e0c09cceb3910f2065 (-`SUSUWU_SCAN_GIT()`, +`SUSUWU_PRODUCTION_USE()`,), which removed `-DSUSUWU_EXPERIMENTAL` from `Macros.sh`.
@`build.sh`, @`compile_commands.json`;
	Set `-DSUSUWU_EXPERIMENTAL -DSUSUWU_DEFAULT_BRANCH=\"trunk\"` (for `git switch <default_branch>` suggestion).
	@`posts/VirusAnalysis.md`; include all of this, + suggest `git switch preview`.

Notice: will consider to [rename this branch (`experimental`) to `preview`](https://github.com/SwuduSusuwu/SusuLib/issues//29).
Warning: This `experimental` branch is for [beta tests](#beta-testexperimental-builds)/[continuous integration](https://google.com?q=continuous-integration-branch).
This is the first commit (limited to this `experimental` branch, which is the branch used to backup&share local (draft) commits.)
The commits in this branch (after this commit): are not official, are not supported, are subject to frequent rebases (so that the `trunk` branch's `git log` is not cluttered with merges/force pushes).
The commits before this commit exist (or soon will) in the `trunk` branch, which must pass all unit tests.
Once a commmit "Q" in the `experimental` branch is fit for use:
	"Q" is moved (`git rebase -i HEAD~N`, where "N" is the count of the list of commits to go back through) before this commit,
	[this commit is now the ancestor of "Q", both now have new hashes]:
		`git switch trunk`,
		`git rebase -i HEAD`, `;%s/noop/pick Q/` #`git cherry-pick Q` changes the commit hash
		`git push origin HEAD:trunk` to the `trunk` branch for all to use.
	or (without switch from `experimental`):
		`git push origin Q:trunk` #commits `Q` to the `trunk` branch (for all to use.)
Thus, this `experimental` branch and the `trunk` branch will have a new common ancestor (which allows the 2 branches to continue as close as possible).

Commits considered for deletion were moved from `experimental` branch into `topic/*` branches;
```
~/SusuLib $ git branch -vv
* (no branch, rebasing experimental)				74e185f @`README.md`; experimental versus trunk branch use
	experimental						e379032 [origin/trunk: ahead 16] @`.ssh/setup.sh`; typo fix (for personal use).
	topic/EXPERIMENTAL_virusAnalysisHook			1f11b6a [origin/trunk: ahead 1] @virusAnalysisHook() +`#ifdef SUSUWU_EXPERIMENTAL`
	topic/performance-inefficient-string-concatenation	c8ac298 @`cxx/AssistantCns.cxx`; `;%s/" + /" += /`,
	topic/usetemplatetemplate				f576d92 [origin/trunk: ahead 1] @`classSysColoredParamStr` Use `template<template<`
	trunk							763bb8a [origin/trunk] @`.gitignore` Update ignore list.
```