+`.github/workflows/code_reviews.yml`
Is followup to commit a2d83217 (-`.github/workflows/codacy.yml`), which removed all (except [CodeQL](https://codeql.github.com/)) reviews.

`--suppress=useStlAlgorithm`; so that source code is more compatible with other languages, is preference to use raw loops.
`--suppress=unreadVariable`; False positive fixes, for code which is on "todo" lists (issues, but not regressions).
`--suppress=noExplicitConstructor`; source code is too verbose with all of those `static_cast`, so preference is implicit constructors.
`--suppress=unmatchedSuppression`; so if this is used with a different version of `cppcheck` (or with some tests disabled), "information: Unmatched suppression: constParameter [unmatchedSuppression]" does not abort the whole run.