@`.github/workflows/code_reviews.yml`;`--suppress`
`--suppress=useStlAlgorithm`; so that source code is more compatible with other languages, is preference to use raw loops.
`--suppress=unreadVariable`; False positive fixes, or code which is on "todo" lists (issues, but not regressions).
`--suppress=noExplicitConstructor`; source code is too verbose with all of those `static_cast`, so preference is implicit constructors.
`--suppress=unmatchedSuppression`; so if this is used with a different version of `cppcheck` (or with some tests disabled), "information: Unmatched suppression: constParameter [unmatchedSuppression]" does not abort the whole run.

Is followup to: commit a690ee75ca38695580649eb21744245a7d3673d6 (+`.github/workflows/code_reviews.yml`). TODO: `squash` this?